Relax Group Replication IP Whitelist
With this patch we restrict to the 10.0.0.0/8 subnet. The default value
is "AUTOMATIC" which results in a /24 net only.

In a default Kuberenetes configuration the Pod's of a single Kubernetes
node form a /24 net, so when having multiple K8s nodes that would be too
restrictive.

We have to check if IP ranges outside the 10.0.0.0/8 range are common
and see if there is a good alternate configuration We can do. In my
quick test 0.0.0.0/0 wa rejected as value. In any case we can augment
the group replication Whitelist with Kuberentes Network Policies to
limit GR communication between a single InnoDB cluster's Pods.