Fixing race condition
The sidecar can be slow with configuring the local to the pod MySQL instance.
In this case the ContainersReady is already set but the sidecar hasn't
yet configured its Server. The operator has to wait for the 'configured'
gate to be set before attempting to configure a cluster with this new
MySQL instance.