Enable stateful set to start in an istio cluster
by making this change i was able to get the cluster to startup within an istio cluster. This isn't the ideal solution as having an Istio virtual service routing to it would be more preferred but this change will at least allow This operator to be used out of the box within an Istio cluster.