Make Router connect to isntance service isntead of specific pod
Previously the routers were configured to alway conenct to Pod -0.
However this would fail if that Pod isn'T available due to an error
situation.

With this change it will connect to the service via
`{clustername}-isntances.svc....` which will resolve to any available
server. The router will then inquire the instances for their proper
names for continued operation.