Fixed permissions of mysql datadir
MySQL datadir was getting created with access settings.
It's now owned by the mysql user and have mode 0700.
A new fixdatadir initContainer runs as root, which fixes
permissions and ownership of the datadir.

Additionally, the mysql user/group is now being created in the
operator image so that the sidecar can run as the mysql user
without issues.

Change-Id: Iba4b3d6821c2646c4161a3d5e5896e55ae25168a