Fix bug #33902183 - creating an instance which can not be provisioned from binlogs fails
With this patch we will always try an incremental approach first and if
Shell fails fall back to clone. This is imperfect as for larger
databases a clone can be more efficient and the retry logic isn't really
nice. This can be improved with Shell's fix for #33900165, which
introduces an "auto" mode.

This also removes incrementalRecoveryAllowed from the cluster_inf, which
is likely wrong for a running system post initial setup.

Change-Id: If92828248fb593472fe7873ec2aa426af43babf3