WL15031 MySQL Operator RBAC Rights
Restrict ClusterRole for the Operator itself and the Role of the operator
to only what is needed and nothing more.
In addition, commented out dev code that granted these to all service accounts
and not only to the operator and sidecar SA.
Fixed how SA is passed to the sidecar. SA can be defined only on top level, as
multiple SA per pod (fine grained for the separate containers) are not supported
natively. For more info see
  https://github.com/kubernetes/kubernetes/issues/66020#issuecomment-590413238