Fix bug#34719171 readinessprobe tries to write without permission
This fix will move the file to /tmp where we have write permissions
for our user as well due to read only root filesystem.

This patch will not update the ConfigMap of existing clusters, thus
only impact new InnoDB Clusters. This should be acceptable since
the impact of the bug is low.

Change-Id: I83667b3986add5e49d08183c87427fa8f07decee