Fix validation of source for spec.router.instances
This field is sourced from value “routerInstances” or
“router.instances”. If “router.instances” is set, we check if it is the
same as “routerInstances”.

We do this with “ne” Helm function. This function cannot compare
different types. This is is quite typical when mixing values from
different sources (for example default values.yaml file and command-line
arguments).

Perhaps more strangely, when values YAML file contains an integer value,
it is treated as float64 by Helm, so even when we use default
values.yaml and set “router.instances” on command line, we get a
strangely looking error message as a result of “printf” trying to print
float64 as int64 (on the lastest version Helm as of this
writing—3.12.1).

We solve both issues by converting both values to strings.