Allow setting spec.router.instances to zero
InnoDBCluster resource can have zero router instances. However, Helm
considers zero to be an empty value (see “default” function†), so we
cannot rely on “coalesce” Helm function to source spec.router.instances
from one of the following values:

1. routerInstances
2. router.instances

We change implementation which relies on Helm functions which use
“emptiness” to the implementation which checks if a value is defined in
parent dict. We do not introduce any additional validation like checking
if the value is an integer zero or any other value or type. This has
been validated by Kubernetes using CRD so far, Helm properly reports
this back, and we do not change this.

† https://helm.sh/docs/chart_template_guide/function_list/#default