Fix bug#3598306 Keyring file and encrypted_file should default to mounted volume
This change primarily fixes that keyring UDF plugin is being loaded on
all pods and keyring will actually be stored on the
volume the user mounted. For ease of use a default file name is also
provided.

Also tests are fixed to actually verify thigns work instead of writing
messages to log which nobody checks ...

Change-Id: Ifb0ace2f9d8b1601f47d3205902463a423caae9f