Fix for Bug#36946536 Containers created by Operator should have explicit nonRoot flag set
It is more informative to the K8s API than making it more secure as the uids
under which the containers run don't change. The K8s will just check them.
The only implication is that if one needs to run an additional container as root
then he needs to change the security ctx of the container to override the global
value of runAsNonRoot.

Change-Id: I0ab11fbf23e9fe3116c7d450f63af80f25569511