Better handling of TLS secrets mouting in the STS.
CA and TLS secrets need to be mounted in different locations, as there could be
overlappingly named elements having different values, which leads to undefined behavior.
This method of using separate directories is already used in the router deployment.
Not mounting just specific files allows for flexibility of adding extra elements to
the secrets, like crls (certificate revocation list), which might or might not be provided.
Inspecting the secrets and mounting only what is in the secret leads to STS rolling update
which is unwanted.

Change-Id: I053856b6ca4befd6d7e9498c8362cf665795f8ef