Make on_pod_event less noisy in the logs. Improved the handler to not busy wait
@kopf.on.event( is too low level and such handlers are not restarted when
they raise KopfTemporaryError. The solution was to busy wait (sleep). The
optimal solution is to use kopf.on.field for watching status.* .
Unfortunately we can't watch just status.containerStatuses, which would
limit the handler to much less events because we need also other status fields.
In any case this will make the handler more stable and less noisy, as it won't
react on every event the pod is going through but only on status changes, where
we can find the restart count that we need

Change-Id: I0d69e11cf97d7973c787c8711f31c0ac0a25c231